backend/requirements/base.txt
Django==5.1.0
djangorestframework==3.15.2
psycopg[binary]==3.3.3
python-decouple==3.8
django-cors-headers==4.4.0
django-filter==24.2
drf-spectacular==0.27.2
django-allauth==0.63.6
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
channels==4.1.0
channels-redis==4.2.0
whitenoise==6.7.0
gunicorn==22.0.0
